#455 refactor interface classes
