+++ pybamm/models/lead_acid/newman_tiedemann.py
@@ -70,7 +70,7 @@ class NewmanTiedemann(pybamm.LeadAcidBaseModel):
         param = self.set_of_parameters
 
         # Exchange-current density
-        int_curr_model = pybamm.interface.LeadAcidReaction(param)
+        int_curr_model = pybamm.interface_lead_acid.MainReaction(param)
         c_e = self.variables["Electrolyte concentration"]
         c_e_n, _, c_e_p = c_e.orphans
