+++ tests/unit/test_models/test_submodels/test_interface.py
@@ -4,7 +4,7 @@ import unittest
 
 class TestInterface(unittest.TestCase):
     def test_domain_failures(self):
-        model = pybamm.interface.InterfacialCurrent(None)
+        model = pybamm.interface.InterfacialReaction(None)
 
         with self.assertRaises(pybamm.DomainError):
             model.get_homogeneous_interfacial_current(None, "not a domain")
@@ -16,13 +16,17 @@ class TestInterface(unittest.TestCase):
             model.get_inverse_butler_volmer(None, None, "not a domain")
 
         c = pybamm.Variable("c", "not a domain")
-        lead_acid_model = pybamm.interface.LeadAcidReaction(None)
-        with self.assertRaises(pybamm.DomainError):
-            lead_acid_model.get_exchange_current_densities(c, "not a domain")
-
         lithium_ion_model = pybamm.interface.LithiumIonReaction(None)
         with self.assertRaises(pybamm.DomainError):
             lithium_ion_model.get_exchange_current_densities(c, "not a domain")
+
+
+class TestInterfaceLeadAcid(unittest.TestCase):
+    def test_domain_failures(self):
+        c = pybamm.Variable("c", "not a domain")
+        lead_acid_model = pybamm.interface_lead_acid.MainReaction(None)
+        with self.assertRaises(pybamm.DomainError):
+            lead_acid_model.get_exchange_current_densities(c, "not a domain")
 
 
 if __name__ == "__main__":
